feat: Move files to have a clearer API
Clearer distinction between MUI components (in Dialog folder) and Cozy
ready-made Dialog components (in CozyDialogs folder).

```patch
- import CozyConfirmDialog from 'cozy-ui/transpiled/react/Dialog/CozyDialogs/CozyConfirmDialog'
+ import { ConfirmDialog } from 'cozy-ui/transpiled/react/CozyDialogs'
```
