+++ react/CozyDialogs/Dialog.jsx
@@ -3,26 +3,30 @@ import PropTypes from 'prop-types'
 import cx from 'classnames'
 
 import { useCozyDialog } from './useCozyDialog'
-import Dialog, {
+import MUIDialog, {
   DialogTitle,
   DialogActions,
-  DialogContent,
-  DialogBackButton,
-  DialogCloseButton
+  DialogContent
 } from 'cozy-ui/transpiled/react/Dialog'
+import { CardDivider } from 'cozy-ui/transpiled/react/MuiCozyTheme/Divider'
+
+import DialogBackButton from './DialogBackButton'
+import DialogCloseButton from './DialogCloseButton'
 import DialogTransition from './DialogTransition'
-const CozyConfirmDialog = ({
+
+const Dialog = ({
   opened,
   onClose,
   title,
   content,
   actions,
-  actionsLayout
+  actionsLayout,
+  size
 }) => {
-  const { cssSize, isFullscreen, id } = useCozyDialog('S')
+  const { cssSize, isFullscreen, id } = useCozyDialog(size)
 
   return (
-    <Dialog
+    <MUIDialog
       open={opened}
       onClose={onClose}
       TransitionComponent={DialogTransition}
@@ -32,38 +36,40 @@ const CozyConfirmDialog = ({
       aria-labelledby={`modal-title-${id}`}
     >
       {!isFullscreen && <DialogCloseButton onClick={onClose} />}
+      <DialogTitle
+        id={`modal-title-${id}`}
+        disableTypography
+        className="u-ellipsis"
+      >
+        {isFullscreen ? <DialogBackButton onClick={onClose} /> : null}
+        {title}
+      </DialogTitle>
+      <CardDivider />
       <DialogContent>
         <div className="dialogContentInner withFluidActions">
-          <DialogTitle
-            id={`modal-title-${id}`}
-            disableTypography
-            className="dialogTitleFluid"
-          >
-            {isFullscreen ? <DialogBackButton onClick={onClose} /> : null}
-            {title}
-          </DialogTitle>
           {content}
           <DialogActions
             disableActionSpacing
             className={cx('dialogActionsFluid', {
-              columnLayout: actionsLayout == "column"
+              columnLayout: actionsLayout == 'column'
             })}
           >
             {actions}
           </DialogActions>
         </div>
       </DialogContent>
-    </Dialog>
+    </MUIDialog>
   )
 }
 
-CozyConfirmDialog.propTypes = {
+Dialog.propTypes = {
   opened: PropTypes.bool.isRequired,
   onClose: PropTypes.func.isRequired,
   title: PropTypes.node,
   content: PropTypes.node,
   actions: PropTypes.node,
-  actionsLayout: PropTypes.oneOf(['row', 'column'])
+  actionsLayout: PropTypes.oneOf(['row', 'column']),
+  size: PropTypes.string
 }
 
-export default CozyConfirmDialog
+export default Dialog
