+++ react/CozyDialogs/DialogCloseButton.jsx
