+++ react/CozyDialogs/index.jsx
@@ -1,0 +1,8 @@
+export { default as DialogBackButton } from './DialogBackButton'
+export { default as DialogCloseButton } from './DialogCloseButton'
+
+export { default as Dialog } from './Dialog'
+export { default as ConfirmDialog } from './ConfirmDialog'
+export { default as FixedActionsDialog } from './FixedActionsDialog'
+export { default as FixedDialog } from './FixedDialog'
+export { default as IllustrationDialog } from './IllustrationDialog'
