+++ react/Dialog/index.jsx
@@ -6,5 +6,3 @@ export { default as DialogActions } from './DialogActions'
 export { default as DialogContent } from './DialogContent'
 export { default as DialogContentText } from './DialogContentText'
 export { default as DialogTitle } from './DialogTitle'
-export { default as DialogBackButton } from './DialogBackButton'
-export { default as DialogCloseButton } from './DialogCloseButton'
